requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
bootstrap4==0.1.0
cssselect==1.1.0
dj-database-url==0.5.0
Django==3.2.4
django-bootstrap4==3.0.1
django-cities==0.6
django-countries==7.2.1
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-language-fields==1.0.0
django-model-utils==4.1.1
django-phone-field==1.8.1
django-phonenumber-field==5.2.0
django-toolbox==0.2
django-world-languages==0.1
djangorestframework==3.12.4
djangotoolbox==1.8.0
gunicorn==20.1.0
lxml==4.6.3
phonenumbers==8.12.28
Pillow==8.3.1
psycopg2==2.9.1
pyquery==1.4.3
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
swapper==1.1.2.post1
tqdm==4.62.0
whitenoise==5.3.0
